% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{'Groq': Support Functions}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{The name of the library where this add-on is located.}

\item{pkgname}{The name of the package that is being attached.}
}
\description{
Runs when attached such as by library() or require()
}
\details{
This function is an attachment point for the 'onAttach' event in R packages. It is called when this package is attached to the R session."
}
