% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groqR-package.R
\docType{package}
\name{groqR-package}
\alias{groqR}
\alias{groqR-package}
\title{groqR: A Coding Assistant using the Groq API}
\description{
A coding assistant using the super fast Groq API \url{https://groq.com/} API. A set of functions and 'RStudio' Addins that aim to help R developers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GabrielKaiserQFin/groqR}
  \item Report bugs at \url{https://github.com/GabrielKaiserQFin/groqR/issues}
}

}
\author{
\strong{Maintainer}: Gabriel Kaiser \email{quantresearch.gk@gmail.com}

}
\keyword{internal}
