% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{modelCall}
\alias{modelCall}
\title{Function to query current models}
\usage{
modelCall(api_key = Sys.getenv("GROQ_API_KEY"))
}
\arguments{
\item{api_key}{The Groq API key.}
}
\value{
model table
}
\description{
The \code{modelCall} function is designed to  query current models from Groq.
}
\details{
The function checks the following environment variables:
\itemize{
\item \code{GROQ_API_KEY}
}
}
