% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{serverInit}
\alias{serverInit}
\title{Set GROQ Parameters Server Logic}
\usage{
serverInit(input, output, session)
}
\arguments{
\item{input}{The input values as they are submitted by the user.}

\item{output}{The output values.}

\item{session}{The Shiny session object.}
}
\description{
This function handles the server logic for the UI, including the observeEvent
for the action button. It asks the user to prompt various GROQ inputs.
Once the 'proxy' parameter is set, the parameters are written to the
environment as environment variables and a notification message is displayed.
}
