% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SupportFcts.R
\docType{data}
\name{ui}
\alias{ui}
\title{Groq: Support Functions}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 4.
}
\usage{
ui
}
\description{
User Interface
}
\details{
It creates a fluidPage with a title and an action button for triggering
the input event.
}
\keyword{datasets}
