% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{uiInit}
\alias{uiInit}
\title{User Interface}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 4.
}
\usage{
uiInit
}
\description{
It creates a fluidPage with a title and an action button for triggering
the input event.
}
\keyword{datasets}
