% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{Update.Renviron}
\alias{Update.Renviron}
\title{Update Renviron file with given list of variables and paths}
\usage{
Update.Renviron(my_list)
}
\arguments{
\item{my_list}{A named list containing variables and their values to be added
or updated in the .Renviron file.}
}
\description{
Update Renviron file with given list of variables and paths
}
