% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groqR-package.R
\docType{package}
\name{groqR-package}
\alias{groqR}
\alias{groqR-package}
\title{groqR: A Coding Assistant using the Fast AI Inference 'Groq'}
\description{
A comprehensive suite of functions and 'RStudio' Add-ins leveraging the capabilities of open-source Large Language Models (LLMs) to support R developers. These functions offer a range of utilities, including text rewriting, translation, and general query capabilities. Additionally, the programming-focused functions provide assistance with debugging, translating, commenting, documenting, and unit testing code, as well as suggesting variable and function names, thereby streamlining the development process.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GabrielKaiserQFin/groqR}
  \item Report bugs at \url{https://github.com/GabrielKaiserQFin/groqR/issues}
}

}
\author{
\strong{Maintainer}: Gabriel Kaiser \email{quantresearch.gk@gmail.com}

}
\keyword{internal}
