% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APIcall.R, R/zzz.R
\name{Update.Renviron}
\alias{Update.Renviron}
\title{Update Renviron file with given list of variables and paths}
\usage{
Update.Renviron(my_list)

Update.Renviron(my_list)
}
\arguments{
\item{my_list}{A named list containing variables and their values to be added
or updated in the .Renviron file.}

\item{renviron_path}{The file path of the existing .Renviron file.
(@code{file.path(Sys.getenv('HOME'), '.Renviron')}) by default.}
}
\description{
Update Renviron file with given list of variables and paths

Update Renviron file with given list of variables and paths
}
